requirements-api.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
psycopg2-binary>=2.9.9
